Add feed

+++ src/main/java/ru/yandex/practicum/filmorate/services/ReviewService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exceptions.EmptyResultFromDataBaseException;
 import ru.yandex.practicum.filmorate.models.Review;
+import ru.yandex.practicum.filmorate.storage.interf.EventStorage;
 import ru.yandex.practicum.filmorate.storage.interf.ReviewStorage;
 
 import java.util.List;
@@ -18,6 +19,8 @@ import java.util.Map;
 public class ReviewService {
     private final ReviewStorage<Review> reviewStorage;
 
+    private final EventStorage eventStorage;
+
     public List<Review> getAll() {
         return reviewStorage.getAll();
     }
@@ -27,11 +30,45 @@ public class ReviewService {
     }
 
     public Review create(Review review) {
-        return reviewStorage.create(review);
+        if (!reviewStorage.checkFilmExists(review.getFilmId())) {
+            throw new EmptyResultFromDataBaseException("Фильм с идентификатором " + review.getFilmId() + " отсутствует");
+        }
+
+        if (!reviewStorage.checkUserExists(review.getUserId())) {
+            throw new EmptyResultFromDataBaseException("Пользователь с идентификатором " + review.getUserId() + " отсутствует");
+        }
+
+        Review createdReview = reviewStorage.create(review);
+
+        Map<String, Object> params = eventStorage.makeEvent(
+                (long)review.getUserId(),
+                Math.toIntExact(createdReview.getReviewId()),
+                "review",
+                "add"
+        );
+        eventStorage.save(params);
+        return createdReview;
     }
 
     public Review update(Review review) {
-        return reviewStorage.update(review);
+        if (!reviewStorage.checkFilmExists(review.getFilmId())) {
+            throw new EmptyResultFromDataBaseException("Фильм с идентификатором " + review.getFilmId() + " отсутствует");
+        }
+        if (!reviewStorage.checkUserExists(review.getUserId())) {
+            throw new EmptyResultFromDataBaseException("Пользователь с идентификатором " + review.getUserId() + " отсутствует");
+        }
+
+        Review updatedReview = reviewStorage.update(review);
+
+        Map<String, Object> params = eventStorage.makeEvent(
+                (long)updatedReview.getUserId(),
+                Math.toIntExact(updatedReview.getReviewId()),
+                "review",
+                "update"
+        );
+        eventStorage.save(params);
+        
+        return updatedReview;
     }
 
     public Review getByReviewId(Long reviewId) {
@@ -39,14 +76,38 @@ public class ReviewService {
     }
 
     public void addLikeToReview(Integer reviewId, Integer userId) {
+        if (!reviewStorage.checkUserExists(userId)) {
+            throw new EmptyResultFromDataBaseException("Пользователь с идентификатором " + userId + " отсутствует");
+        }
+
+        if (!reviewStorage.checkReviewExists(reviewId)) {
+            throw new EmptyResultFromDataBaseException("Отзыв с идентификатором " + reviewId + " отсутствует");
+        }
         reviewStorage.addLikeToReview(reviewId, userId);
     }
 
     public void addDisLikeToReview(Integer reviewId, Integer userId) {
+        if (!reviewStorage.checkUserExists(userId)) {
+            throw new EmptyResultFromDataBaseException("Пользователь с идентификатором " + userId + " отсутствует");
+        }
+
+        if (!reviewStorage.checkReviewExists(reviewId)) {
+            throw new EmptyResultFromDataBaseException("Отзыв с идентификатором " + reviewId + " отсутствует");
+        }
         reviewStorage.addDisLikeToReview(reviewId, userId);
     }
 
     public void delete(Long reviewId) {
+        if (!reviewStorage.checkReviewExists(Math.toIntExact(reviewId))) {
+            throw new EmptyResultFromDataBaseException("Отзыв с идентификатором " + reviewId + " отсутствует");
+        }
+        Map<String, Object> params = eventStorage.makeEvent(
+                1L,
+                Math.toIntExact(reviewId),
+                "review",
+                "remove"
+        );
+        eventStorage.save(params);
         reviewStorage.delete(reviewId);
     }
 }
